# Support complex types in sparksql hash and xxhash64 function