--- a/velox/functions/sparksql/tests/HashTest.cpp
+++ b/velox/functions/sparksql/tests/HashTest.cpp
@@ -18,6 +18,8 @@
 
 #include <stdint.h>
 
+using facebook::velox::test::assertEqualVectors;
+
 namespace facebook::velox::functions::sparksql::test {
 namespace {
 
@@ -26,6 +28,10 @@ class HashTest : public SparkFunctionBaseTest {
   template <typename T>
   std::optional<int32_t> hash(std::optional<T> arg) {
     return evaluateOnce<int32_t>("hash(c0)", arg);
+  }
+
+  VectorPtr hash(VectorPtr vector) {
+    return evaluate("hash(c0)", makeRowVector({vector}));
   }
 };
 
@@ -126,6 +132,106 @@ TEST_F(HashTest, Float) {
   EXPECT_EQ(hash<float>(limits::quiet_NaN()), -349261430);
   EXPECT_EQ(hash<float>(limits::infinity()), 2026854605);
   EXPECT_EQ(hash<float>(-limits::infinity()), 427440766);
+}
+
+TEST_F(HashTest, Array) {
+  assertEqualVectors(
+      makeFlatVector<int32_t>({2101165938, 42, 1045631400}),
+      hash(makeArrayVector<int64_t>({{1, 2, 3, 4, 5}, {}, {1, 2, 3}})));
+
+  assertEqualVectors(
+      makeFlatVector<int32_t>({-559580957, 1765031574, 42}),
+      hash(makeNullableArrayVector<int32_t>(
+          {{1, std::nullopt}, {std::nullopt, 2}, {std::nullopt}})));
+
+  // Nested array.
+  {
+    using innerArrayType = std::vector<std::optional<int64_t>>;
+    using outerArrayType =
+        std::vector<std::optional<std::vector<std::optional<int64_t>>>>;
+
+    innerArrayType a{1, std::nullopt, 2, 3};
+    innerArrayType b{4, 5};
+    innerArrayType c{6, 7, 8};
+    outerArrayType row1{{a}, {b}};
+    outerArrayType row2{{a}, {c}};
+    outerArrayType row3{{{}}};
+    outerArrayType row4{{{std::nullopt}}};
+    auto arrayVector = makeNullableNestedArrayVector<int64_t>(
+        {{row1}, {row2}, {row3}, {row4}, std::nullopt});
+    assertEqualVectors(
+        makeFlatVector<int32_t>({2101165938, -992561130, 42, 42, 42}),
+        hash(arrayVector));
+  }
+
+  // Array of map.
+  {
+    using S = StringView;
+    using P = std::pair<int64_t, std::optional<S>>;
+    std::vector<P> a{P{1, S{"a"}}, P{2, std::nullopt}};
+    std::vector<P> b{P{3, S{"c"}}};
+    std::vector<std::vector<std::vector<P>>> data = {{a, b}};
+    auto arrayVector = makeArrayOfMapVector<int64_t, S>(data);
+    assertEqualVectors(
+        makeFlatVector<int32_t>(std::vector<int32_t>{-718462205}),
+        hash(arrayVector));
+  }
+
+  // Array of row.
+  {
+    std::vector<std::vector<std::optional<std::tuple<int32_t, std::string>>>>
+        data = {
+            {{{1, "red"}}, {{2, "blue"}}, {{3, "green"}}},
+            {{{1, "red"}}, std::nullopt, {{3, "green"}}},
+            {std::nullopt},
+        };
+    auto arrayVector = makeArrayOfRowVector(data, ROW({INTEGER(), VARCHAR()}));
+    assertEqualVectors(
+        makeFlatVector<int32_t>({-1458343314, 551500425, 42}),
+        hash(arrayVector));
+  }
+}
+
+TEST_F(HashTest, Map) {
+  auto mapVector = makeMapVector<int64_t, double>(
+      {{{1, 17.0}, {2, 36.0}, {3, 8.0}, {4, 28.0}, {5, 24.0}, {6, 32.0}}});
+  assertEqualVectors(
+      makeFlatVector<int32_t>(std::vector<int32_t>{1263683448}),
+      hash(mapVector));
+
+  auto mapOfArrays = createMapOfArraysVector<int32_t, int32_t>(
+      {{{1, {{1, 2, 3}}}}, {{2, {{4, 5, 6}}}}, {{3, {{7, 8, 9}}}}});
+  assertEqualVectors(
+      makeFlatVector<int32_t>({-1818148947, 529298908, 825098912}),
+      hash(mapOfArrays));
+
+  auto mapWithNullArrays = createMapOfArraysVector<int64_t, int64_t>(
+      {{{1, std::nullopt}}, {{2, {{4, 5, std::nullopt}}}}, {{3, {{}}}}});
+  assertEqualVectors(
+      makeFlatVector<int32_t>({-1712319331, 2060637564, 519220707}),
+      hash(mapWithNullArrays));
+}
+
+TEST_F(HashTest, Row) {
+  auto row = makeRowVector({
+      makeFlatVector<int64_t>({1, 3}),
+      makeFlatVector<int64_t>({2, 4}),
+  });
+  assertEqualVectors(
+      makeFlatVector<int32_t>({-1181176833, 1717636039}), hash(row));
+
+  row = makeRowVector({
+      makeNullableFlatVector<int64_t>({1, std::nullopt}),
+      makeNullableFlatVector<int64_t>({std::nullopt, 4}),
+  });
+  assertEqualVectors(
+      makeFlatVector<int32_t>({-1712319331, 1344313940}), hash(row));
+
+  row->setNull(0, true);
+  assertEqualVectors(makeFlatVector<int32_t>({42, 1344313940}), hash(row));
+
+  row->setNull(1, true);
+  assertEqualVectors(makeFlatVector<int32_t>({42, 42}), hash(row));
 }
 
 } // namespace