… velox/docs/functions/spark/binary.rst