@jinchengchenghh 2024-04-15
Please update the PR title to note support complex type in `hash and xxhash` function