@marin-ma 2024-04-15
> Is it possible to reconstruct code to avoid too much static functions?

Curious any downside of using static functions? The hash computations are stateless.