@FelixYBW 2024-04-19
@mbasmanova Any more comments? The function is used by Gluten columnar shuffle.