@marin-ma 2024-05-09
@mbasmanova @pedroerp I created a benchmark in https://github.com/facebookincubator/velox/pull/9414/commits/9265b975152586ebdda5576314133a3c24120f53

Benchmark Result at commit https://github.com/facebookincubator/velox/pull/9414/commits/569500df1ffb26642c44233af75941561cd10c13
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
hash_ARRAY<MAP<INTEGER,VARCHAR>>##hash                     15.02ms     66.60
hash_ARRAY<MAP<INTEGER,VARCHAR>>##xxhash64                 14.27ms     70.05
hash_ROW<f_map:MAP<INTEGER,VARCHAR>,f_array:ARR            25.94ms     38.55
hash_ROW<f_map:MAP<INTEGER,VARCHAR>,f_array:ARR            25.99ms     38.48
```

Benchmark Result at commit https://github.com/facebookincubator/velox/pull/9414/commits/9265b975152586ebdda5576314133a3c24120f53 (latest)
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
hash_ARRAY<MAP<INTEGER,VARCHAR>>##hash                     17.76ms     56.31
hash_ARRAY<MAP<INTEGER,VARCHAR>>##xxhash64                 16.20ms     61.71
hash_ROW<f_map:MAP<INTEGER,VARCHAR>,f_array:ARR            28.24ms     35.41
hash_ROW<f_map:MAP<INTEGER,VARCHAR>,f_array:ARR            28.32ms     35.31
```

Seems like the previous implementation at https://github.com/facebookincubator/velox/pull/9414/commits/569500df1ffb26642c44233af75941561cd10c13 is better.