@FelixYBW 2024-05-09
> @mbasmanova @pedroerp I created a benchmark in [9265b97](https://github.com/facebookincubator/velox/commit/9265b975152586ebdda5576314133a3c24120f53)
> 
> Benchmark Result at commit [569500d](https://github.com/facebookincubator/velox/commit/569500df1ffb26642c44233af75941561cd10c13)
> 
> ```
> ============================================================================
> [...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
> ============================================================================
> hash_ARRAY<MAP<INTEGER,VARCHAR>>##hash                     15.02ms     66.60
> hash_ARRAY<MAP<INTEGER,VARCHAR>>##xxhash64                 14.27ms     70.05
> hash_ROW<f_map:MAP<INTEGER,VARCHAR>,f_array:ARR            25.94ms     38.55
> hash_ROW<f_map:MAP<INTEGER,VARCHAR>,f_array:ARR            25.99ms     38.48
> ```
> 
> Benchmark Result at commit [9265b97](https://github.com/facebookincubator/velox/commit/9265b975152586ebdda5576314133a3c24120f53) (latest)
> 
> ```
> ============================================================================
> [...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
> ============================================================================
> hash_ARRAY<MAP<INTEGER,VARCHAR>>##hash                     17.76ms     56.31
> hash_ARRAY<MAP<INTEGER,VARCHAR>>##xxhash64                 16.20ms     61.71
> hash_ROW<f_map:MAP<INTEGER,VARCHAR>,f_array:ARR            28.24ms     35.41
> hash_ROW<f_map:MAP<INTEGER,VARCHAR>,f_array:ARR            28.32ms     35.31
> ```
> 
> Seems like the previous implementation at [569500d](https://github.com/facebookincubator/velox/commit/569500df1ffb26642c44233af75941561cd10c13) is better.

I didn't go through your code. If I understand correctly, it's the comparison between an indirect branch prediction vs. direct branch prediction.

If the prediction is 100% random, indirect branch should be better. If it's 100% predictable, direct one is better.