@FelixYBW 2024-05-09
> 
> I didn't go through your code. If I understand correctly, it's the comparison between an indirect branch prediction vs. direct branch prediction.
> 
> If the prediction is 100% random, indirect branch should be better. If it's 100% predictable, direct one is better.

@marin-ma just for sure, let's collect some PMU events to confirm.