@marin-ma 2024-05-10
> Would you confirm that result are from running benchmark in 'release' mode?

@mbasmanova Yes, in 'release' mode.