@FelixYBW 2024-05-10
@mbasmanova , @marin-ma collected below data. conclusion is to use the virtual function call.

Virtual function call has 15% better performance. The root cause is it has 20% less instructions. It's expected the switch statement has more branch instructions.
But in both case, the branch misprediction aren't a issue, miss ratio is 2.8%. Compared to Virtual function call, the switch method doesn't show higher branch misprediction because the switch in the function has fixed patten which comes from the schema, it can be very complex but BPU can handle this well as long as it has fixed patten. The misprediction ratio may increase if we have a really complex schema which exceeds BPU's history track buffer which is in 1000 level now.



per second | virtual function call | switch |  
-- | -- | -- | --
INST_RETIRED.ANY | 7,850,078,349 | 8,586,837,809 |  
CPU_CLK_UNHALTED.REF_TSC | 2,600,825,704 | 2,600,865,762 |  
CPU_CLK_UNHALTED.THREAD | 3,094,305,687 | 3,094,298,604 |  
CYCLE_ACTIVITY.STALLS_L3_MISS | 694,925 | 915,935 |  
LONGEST_LAT_CACHE.MISS | 261,666 | 340,501 |  
MEM_LOAD_RETIRED.L3_MISS | 14,705 | 17,211 |  
BR_INST_RETIRED.ALL_BRANCHES | 1,087,985,060 | 1,400,282,806 |  
BR_MISP_RETIRED.ALL_BRANCHES | 30,420,663 | 27,806,660 |  
BR_INST_RETIRED.INDIRECT | 93,107,159 | 39,964,183 |  
BR_MISP_RETIRED.INDIRECT | 1,900 | 145,261 |  
  |   |   | virtual/switch
Instruction   per loop | 14,714 | 18,584 | 0.792
IPC | 2.54 | 2.78 |  
loop per   second | 533,513 | 462,064 | 1.155
branch   misprediction ratio | 2.8% | 2.0%
branch   misprediction/1K inst | 3.88 | 3.24
branch per loop | 2,039 | 3,030