@FelixYBW 2024-05-11
@marin-ma Let's use virtual function way for now. Can you update the PR?