@marin-ma 2024-05-11
@mbasmanova Reverted to virtual function implementation. Looks like the UT failure is not caused by this change. Could you help to review again? Thanks!