@mbasmanova 2024-05-15
@FelixYBW @marin-ma Folks, I used your benchmark to compare performance before and after this PR for primitive types. I'm seeing a significant regression.

Before:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
hash_BIGINT##hash                                         585.43us     1.71K
hash_BIGINT##xxhash64                                     488.31us     2.05K
----------------------------------------------------------------------------
hash_INTEGER##hash                                        488.94us     2.05K
hash_INTEGER##xxhash64                                    471.50us     2.12K
hash_VARCHAR##hash                                          2.21ms    452.07
hash_VARCHAR##xxhash64                                      1.97ms    508.80
```

After:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
hash_BIGINT##hash                                         883.61us     1.13K
hash_BIGINT##xxhash64                                     927.77us     1.08K
----------------------------------------------------------------------------
hash_INTEGER##hash                                        808.87us     1.24K
hash_INTEGER##xxhash64                                    815.35us     1.23K
hash_VARCHAR##hash                                          2.67ms    374.56
hash_VARCHAR##xxhash64                                      2.49ms    402.09
```